Refactor: change models to own file
