+++ src/main/java/com/tenniscourts/exceptions/ErrorDetails.java
