+++ src/main/java/com/tenniscourts/reservations/model/ReservationStatus.java
@@ -1,4 +1,4 @@
-package com.tenniscourts.reservations;
+package com.tenniscourts.reservations.model;
 
 public enum ReservationStatus {
   READY_TO_PLAY,
