+++ src/main/java/com/tenniscourts/schedules/ScheduleController.java
@@ -1,6 +1,8 @@
 package com.tenniscourts.schedules;
 
 import com.tenniscourts.config.BaseRestController;
+import com.tenniscourts.schedules.model.CreateScheduleRequestDTO;
+import com.tenniscourts.schedules.model.ScheduleDTO;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
