+++ src/main/java/com/tenniscourts/tenniscourts/TennisCourtRepository.java
@@ -1,5 +1,6 @@
 package com.tenniscourts.tenniscourts;
 
+import com.tenniscourts.tenniscourts.model.TennisCourt;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TennisCourtRepository extends JpaRepository<TennisCourt, Long> {
